Crate hiirc

hiirc is a library built on top of loirc. The goal is to offer a friendly, event-based API.

To use the library, implement the listener trait and give an instance of your implementation to the dispatch function, accompanied with an instance of the Settings struct configured to your needs.

Modules

ext
Utilities that are not part of the official IRC protocol standard.

Structs

Channel
Represents a channel.

ChannelUser
Represents a user inside of a channel.

Irc
Represents the state of this connection.

Message
Represents a message received from the server.

MonitorSettings
These settings tell the monitor how to behave.

PrefixUser
User prefix representation.

Settings
Settings for the dispatcher.

Enums

ChannelUserStatus
Represents the status of a user inside of a channel.

Code
Representation of IRC commands, replies and errors.

ConnectionStatus
Status of the connection.

Error
Errors that can occur.

Event
This is the comprehensive set of events that can occur.

LoircError
Errors produced by the Writer.

ParseError
Error generated by the parser.

Prefix
Prefix of the message.

ReconnectionSettings
These settings tell the reconnection process how to behave.

Traits

Listener
Implement this trait to handle events.

Functions

dispatch
Create an irc client with the listener and settings.