Enum hiirc::Event

pub enum Event {
    Closed(&'static str),
    Disconnected,
    Message(Message),
    ParseError(ParseError),
    Reconnected,
    Reconnecting,
    ReconnectionError(Error),
}

This is the comprehensive set of events that can occur.

Variants

Closed

Connection was manually closed. The string is the reason.

Disconnected

Connection has dropped.

Message

Message from the IRC server.

ParseError

Error parsing a message from the server.

This can probably be ignored, and it shouldn't ever happen, really. If you catch this you should probably open an issue on GitHub.

Reconnected

Connection was sucessfully restored.

Reconnecting

Attempting to restore connection.

ReconnectionError

An error occured trying to restore the connection.

This is normal in poor network conditions. It might take a few attempts before the connection can be restored.