Enum hiirc::ReconnectionSettings

pub enum ReconnectionSettings {
    DoNotReconnect,
    pub Reconnect {
        pub max_attempts: u32,
        pub delay_between_attempts: Duration,
        pub delay_after_disconnect: Duration,
    },
}

These settings tell the reconnection process how to behave.

Default is implemented for this type, with fairly sensible settings. See the Default trait implementation.

Variants

DoNotReconnect

Don't try to reconnect after failure.

Reconnect

Reconnect

Fields

max_attempts

delay_between_attempts

delay_after_disconnect