Struct hiirc::ChannelUser

pub struct ChannelUser {
    // some fields omitted
}

User inside a channel.

Note that the same person might be in many channels. In any case, there will be a ChannelUser object for each Channel the person is in.

Methods

impl ChannelUser

fn nickname(&self) -> Arc<String>

Get the nickname of the user.

fn status(&self) -> ChannelUserStatus

Get the status of the user.