Struct hiirc::DeferredWriter

pub struct DeferredWriter {
    // some fields omitted
}

The DeferredWriter can send commands to the irc server.

The difference with the Irc struct is that the DeferredWriter contains no data about the state and channels. This means that it can be cloned without any fear and used at any time. This is useful when calling methods that might block for a while. You can call them in a seperate thread and not block the event loop.