Struct hiirc::MonitorSettings

pub struct MonitorSettings {
    pub activity_timeout: Duration,
    pub ping_timeout: Duration,
}

These settings tell the monitor how to behave.

They allow you to configure the amount of time between the steps.

Fields
activity_timeout: Duration

Amount of time since the last activity.

When the amount of time since the last activity gets higher than this value, tt will trigger a ping request. If there is not a lot of activity, this means that the server will be pinged everytime this duration expires.

ping_timeout: Duration

Amount of time to wait for a ping reply.

When the amount of time since the ping was sent gets higher than this value, and that no activity occured, assume the connection was dropped and trigger and a disconnect.