Trait hiirc::Listener

pub trait Listener {
    fn any(&mut self, irc: &Irc, event: &Event) { ... }
    fn welcome(&mut self, irc: &Irc) { ... }
    fn channel_join(&mut self, irc: &Irc, channel: &str) { ... }
    fn user_join(&mut self, irc: &Irc, channel: &str, nickname: &str) { ... }
    fn user_part(&mut self, irc: &Irc, channel: &str, nickname: &str) { ... }
    fn user_quit(&mut self, irc: &Irc, nickname: &str) { ... }
    fn channel_msg(&mut self, irc: &Irc, sender: &str, channel: &str, message: &str) { ... }
    fn private_msg(&mut self, irc: &Irc, sender: &str, message: &str) { ... }
    fn topic(&mut self, irc: &Irc, channel: &Channel, topic: Option<&str>) { ... }
    fn topic_change(&mut self, irc: &Irc, channel: &Channel, topic: Option<&str>) { ... }
}

Implement this trait to handle events.

Provided Methods

fn any(&mut self, irc: &Irc, event: &Event)

Any event.

fn welcome(&mut self, irc: &Irc)

When the server sends the welcome packet.

fn channel_join(&mut self, irc: &Irc, channel: &str)

When the client sucessfully joins a channel.

fn user_join(&mut self, irc: &Irc, channel: &str, nickname: &str)

When a user joins a channel we are listening on.

fn user_part(&mut self, irc: &Irc, channel: &str, nickname: &str)

When a user parts a channel we are listening on.

fn user_quit(&mut self, irc: &Irc, nickname: &str)

When a user quits.

fn channel_msg(&mut self, irc: &Irc, sender: &str, channel: &str, message: &str)

When a channel message is received.

fn private_msg(&mut self, irc: &Irc, sender: &str, message: &str)

When a private message is received.

fn topic(&mut self, irc: &Irc, channel: &Channel, topic: Option<&str>)

Reply to a topic command or when joining a channel.

Note that this event might be called before the channel's user list is populated. If a channel has no topic, this event will not be fired when you join a channel. It's safe to assume that a channel has no topic if this event is not fired when joining. If you use the get_topic method however, you will always get a result.

fn topic_change(&mut self, irc: &Irc, channel: &Channel, topic: Option<&str>)

When the topic of is changed by someone.

If you use the set_topic method, you will get a topic_change event instead of a topic event.